fix: move component into correct folder
