+++ src/components/ProjectPicker/project-picker.scss
